+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `approvalsService.closeFormApprovalFlowInstance()`
+- `approvalsService.reopenFormApprovalFlowInstance()`
+
+### Deprecated
+
+- `approvalsService.reopenFormSubmissionApproval()`
 
 ## [2.0.0] - 2022-08-28